+++ src/main.rs
@@ -4,6 +4,7 @@ use std::process::ExitCode;
 
 mod local_rules;
 mod vsg_cli;
+mod waivers;
 
 fn main() -> ExitCode {
     vsg_cli::main(&std::env::args().collect::<Vec<_>>())